requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
tavily-python>=0.7.23
python-dotenv
